Add SSM patch endpoint
This includes a major refactor of the exiting endpoints and code
organization of the SSM endpoints, in order to re-use some code
for SSM operations in the endpoint (by putting them in a base class),
and to prevent the old file project_ssm_configuration.py from
getting too large.

This organizes ssm into its own package with a separate file for
the pydantic models that validate the PATCH request, and the
handlers that get exported in the package.

The aws client code has a minor update to the signature of
put_parameter, since we need to include the parameter type when
creating a new param, but not when overwriting an existing one.
