+++ imbi/clients/aws.py
@@ -47,17 +47,22 @@ async def put_parameter(session: aioboto3.Session,
                         secret_access_key: str,
                         session_token: str,
                         path: str,
-                        param_type: str,
                         value: str,
+                        param_type: str = None,
                         overwrite: bool = False):
+    kwargs = {
+        'Name': path,
+        'Value': value,
+        'Overwrite': overwrite,
+    }
+    if param_type is not None:
+        kwargs['Type'] = param_type
+
     async with session.client('ssm',
                               aws_access_key_id=access_key_id,
                               aws_secret_access_key=secret_access_key,
                               aws_session_token=session_token) as client:
-        await client.put_parameter(Name=path,
-                                   Value=value,
-                                   Type=param_type,
-                                   Overwrite=overwrite)
+        await client.put_parameter(**kwargs)
 
 
 async def delete_parameter(session: aioboto3.Session, access_key_id: str,
